separate map from home widget
